+++ src/Tesseract.Tests/BaseApiTests.cs
@@ -7,12 +7,11 @@ namespace Tesseract.Tests
     [TestFixture]
     public class BaseApiTests
     {
-        [Test,
-         Ignore("GetVersion doesn't work for x64, probably compilation related.")]
-        public void GetVersion_Is304()
+        [Test]
+        public void CanGetVersion()
         {
-            var version = Interop.TessApi.Native.GetVersion();
-            Assert.That(version, Is.EqualTo("3.04"));
+            var version = Interop.TessApi.BaseApiGetVersion();
+            Assert.That(version, Does.StartWith("4.1.1"));
         }
     }
 }