+++ src/Tesseract/Interop/BaseApi.cs
@@ -41,8 +41,8 @@ namespace Tesseract.Interop
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIDelete")]
         void BaseApiDelete(HandleRef ptr);
 
-        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIDetectOS")]
-        int BaseAPIDetectOS(HandleRef handle, ref OSResult result);
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIDetectOrientationScript")]
+        int TessBaseAPIDetectOrientationScript(HandleRef handle, out int orient_deg, out float orient_conf, out IntPtr script_name, out float script_conf);
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetBoolVariable")]
         int BaseApiGetBoolVariable(HandleRef handle, string name, out int value);
@@ -51,7 +51,25 @@ namespace Tesseract.Interop
         int BaseApiGetDoubleVariable(HandleRef handle, string name, out double value);
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetHOCRText")]
-        IntPtr BaseAPIGetHOCRTextInternal(HandleRef handle, int pageNum);
+        IntPtr BaseApiGetHOCRTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetAltoText")]
+        IntPtr BaseApiGetAltoTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetTsvText")]
+        IntPtr BaseApiGetTsvTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetBoxText")]
+        IntPtr BaseApiGetBoxTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetLSTMBoxText")]
+        IntPtr BaseApiGetLSTMBoxTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetWordStrBoxText")]
+        IntPtr BaseApiGetWordStrBoxTextInternal(HandleRef handle, int pageNum);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetUNLVText")]
+        IntPtr BaseApiGetUNLVTextInternal(HandleRef handle);
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetIntVariable")]
         int BaseApiGetIntVariable(HandleRef handle, string name, out int value);
@@ -67,16 +85,23 @@ namespace Tesseract.Interop
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetThresholdedImage")]
         IntPtr BaseAPIGetThresholdedImage(HandleRef handle);
-        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIProcessPages")]
-        int BaseAPIProcessPages(HandleRef handle, string filename, string retry_config, int timeout_millisec, HandleRef renderer);
-        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIProcessPage")]
-        int BaseAPIProcessPage(HandleRef handle, Pix pix, int page_index, string filename, string retry_config, int timeout_millisec, HandleRef renderer);
+
+        // The following were causing issues on Linux/MacOsX when used in .net core
+        //[RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIProcessPages")]
+        //int BaseAPIProcessPages(HandleRef handle, string filename, string retry_config, int timeout_millisec, HandleRef renderer);
+
+        //[RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIProcessPage")]
+        //int BaseAPIProcessPage(HandleRef handle, Pix pix, int page_index, string filename, string retry_config, int timeout_millisec, HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPISetInputName")]
         void BaseAPISetInputName(HandleRef handle, string name);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetDatapath")]
         string BaseAPIGetDatapath(HandleRef handle);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPISetOutputName")]
         void BaseAPISetOutputName(HandleRef handle, string name);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIGetUTF8Text")]
         IntPtr BaseAPIGetUTF8TextInternal(HandleRef handle);
 
@@ -125,7 +150,7 @@ namespace Tesseract.Interop
 
         // Helper functions
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessVersion")]
-        string GetVersion();
+        IntPtr GetVersion();
 
         // result iterator
 
@@ -175,6 +200,27 @@ namespace Tesseract.Interop
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorConfidence")]
         float ResultIteratorGetConfidence(HandleRef handle, PageIteratorLevel level);
 
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorWordFontAttributes")]
+        IntPtr ResultIteratorWordFontAttributes(HandleRef handle, out bool isBold, out bool isItalic, out bool isUnderlined, out bool isMonospace, out bool isSerif, out bool isSmallCaps, out int pointSize, out int fontId);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorWordIsFromDictionary")]
+        bool ResultIteratorWordIsFromDictionary(HandleRef handle);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorWordIsNumeric")]
+        bool ResultIteratorWordIsNumeric(HandleRef handle);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorWordRecognitionLanguage")]
+        IntPtr ResultIteratorWordRecognitionLanguageInternal(HandleRef handle);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorSymbolIsSuperscript")]
+        bool ResultIteratorSymbolIsSuperscript(HandleRef handle);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorSymbolIsSubscript")]
+        bool ResultIteratorSymbolIsSubscript(HandleRef handle);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorSymbolIsDropcap")]
+        bool ResultIteratorSymbolIsDropcap(HandleRef handle);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultIteratorGetPageIterator")]
         IntPtr ResultIteratorGetPageIterator(HandleRef handle);
 
@@ -222,44 +268,71 @@ namespace Tesseract.Interop
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessChoiceIteratorConfidence")]
         float ChoiceIteratorGetConfidence(HandleRef handle);
 
-        #endregion
+        #endregion Choice Iterator
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBaseAPIPrintVariablesToFile")]
-        int BaseApiPrintVariablesToFile(HandleRef handle, string filename); 
+        int BaseApiPrintVariablesToFile(HandleRef handle, string filename);
 
         #region Renderer API
 
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessTextRendererCreate")]
         IntPtr TextRendererCreate(string outputbase);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessAltoRendererCreate")]
+        IntPtr AltoRendererCreate(string outputbase);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessTsvRendererCreate")]
+        IntPtr TsvRendererCreate(string outputbase);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessHOcrRendererCreate")]
         IntPtr HOcrRendererCreate(string outputbase);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessHOcrRendererCreate2")]
         IntPtr HOcrRendererCreate2(string outputbase, int font_info);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessPDFRendererCreate")]
-        IntPtr PDFRendererCreate(string outputbase, string datadir);
+        IntPtr PDFRendererCreate(string outputbase, IntPtr datadir, int textonly);
+        
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessUnlvRendererCreate")]
         IntPtr UnlvRendererCreate(string outputbase);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessBoxTextRendererCreate")]
         IntPtr BoxTextRendererCreate(string outputbase);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessLSTMBoxRendererCreate")]
+        IntPtr LSTMBoxRendererCreate(string outputbase);
+
+        [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessWordStrBoxRendererCreate")]
+        IntPtr WordStrBoxRendererCreate(string outputbase);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessDeleteResultRenderer")]
         void DeleteResultRenderer(HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererInsert")]
         void ResultRendererInsert(HandleRef renderer, HandleRef next);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererNext")]
         IntPtr ResultRendererNext(HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererBeginDocument")]
-        int ResultRendererBeginDocument(HandleRef renderer, string title);
+        int ResultRendererBeginDocument(HandleRef renderer, IntPtr titlePtr);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererAddImage")]
         int ResultRendererAddImage(HandleRef renderer, HandleRef api);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererEndDocument")]
         int ResultRendererEndDocument(HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererExtention")]
         IntPtr ResultRendererExtention(HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererTitle")]
         IntPtr ResultRendererTitle(HandleRef renderer);
+
         [RuntimeDllImport(Constants.TesseractDllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "TessResultRendererImageNum")]
         int ResultRendererImageNum(HandleRef renderer);
-        #endregion
+
+        #endregion Renderer API
     }
 
     internal static class TessApi
@@ -278,6 +351,7 @@ namespace Tesseract.Interop
             + " ocr_line ocrx_word"
             + "'/>\n"
             + "</head>\n<body>\n";
+
         //XHTML End Tag:
         public const string xhtmlEndTag = " </body>\n</html>\n";
 
@@ -303,17 +377,26 @@ namespace Tesseract.Interop
             }
         }
 
+        public static string BaseApiGetVersion()
+        {
+            IntPtr versionHandle = Native.GetVersion();
+            if (versionHandle != IntPtr.Zero)
+            {
+                var result = MarshalHelper.PtrToString(versionHandle, Encoding.UTF8);
+                return result;
+            }
+
+            return null;
+        }
+
         public static string BaseAPIGetHOCRText(HandleRef handle, int pageNum)
         {
-            IntPtr txtHandle = Native.BaseAPIGetHOCRTextInternal(handle, pageNum);
-            if (txtHandle != IntPtr.Zero)
-            {
+            IntPtr txtHandle = Native.BaseApiGetHOCRTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero) {
                 var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
                 TessApi.Native.DeleteText(txtHandle);
                 return htmlBeginTag + result + htmlEndTag;
-            }
-            else
-            {
+            } else {
                 return null;
             }
         }
@@ -321,12 +404,93 @@ namespace Tesseract.Interop
         //Just Copied:
         public static string BaseAPIGetHOCRText2(HandleRef handle, int pageNum)
         {
-            IntPtr txtHandle = Native.BaseAPIGetHOCRTextInternal(handle, pageNum);
+            IntPtr txtHandle = Native.BaseApiGetHOCRTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero) {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return xhtmlBeginTag + result + xhtmlEndTag;
+            } else {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetAltoText(HandleRef handle, int pageNum)
+        {
+            IntPtr txtHandle = Native.BaseApiGetAltoTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero) {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return result;
+            } else {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetTsvText(HandleRef handle, int pageNum)
+        {
+            IntPtr txtHandle = Native.BaseApiGetTsvTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero) {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return result;
+            } else {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetBoxText(HandleRef handle, int pageNum)
+        {
+            IntPtr txtHandle = Native.BaseApiGetBoxTextInternal(handle, pageNum);
             if (txtHandle != IntPtr.Zero)
             {
                 var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
                 TessApi.Native.DeleteText(txtHandle);
-                return xhtmlBeginTag + result + xhtmlEndTag;
+                return result;
+            }
+            else
+            {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetLSTMBoxText(HandleRef handle, int pageNum)
+        {
+            IntPtr txtHandle = Native.BaseApiGetLSTMBoxTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero)
+            {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return result;
+            }
+            else
+            {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetWordStrBoxText(HandleRef handle, int pageNum)
+        {
+            IntPtr txtHandle = Native.BaseApiGetWordStrBoxTextInternal(handle, pageNum);
+            if (txtHandle != IntPtr.Zero)
+            {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return result;
+            }
+            else
+            {
+                return null;
+            }
+        }
+
+        public static string BaseAPIGetUNLVText(HandleRef handle)
+        {
+            IntPtr txtHandle = Native.BaseApiGetUNLVTextInternal(handle);
+            if (txtHandle != IntPtr.Zero)
+            {
+                var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
+                TessApi.Native.DeleteText(txtHandle);
+                return result;
             }
             else
             {
@@ -337,12 +501,9 @@ namespace Tesseract.Interop
         public static string BaseApiGetStringVariable(HandleRef handle, string name)
         {
             var resultHandle = Native.BaseApiGetStringVariableInternal(handle, name);
-            if (resultHandle != IntPtr.Zero)
-            {
+            if (resultHandle != IntPtr.Zero) {
                 return MarshalHelper.PtrToString(resultHandle, Encoding.UTF8);
-            }
-            else
-            {
+            } else {
                 return null;
             }
         }
@@ -350,14 +511,11 @@ namespace Tesseract.Interop
         public static string BaseAPIGetUTF8Text(HandleRef handle)
         {
             IntPtr txtHandle = Native.BaseAPIGetUTF8TextInternal(handle);
-            if (txtHandle != IntPtr.Zero)
-            {
+            if (txtHandle != IntPtr.Zero) {
                 var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
                 TessApi.Native.DeleteText(txtHandle);
                 return result;
-            }
-            else
-            {
+            } else {
                 return null;
             }
         }
@@ -374,19 +532,15 @@ namespace Tesseract.Interop
             string[] varNames = new string[initialValues.Count];
             string[] varValues = new string[initialValues.Count];
             int i = 0;
-            foreach (var pair in initialValues)
-            {
+            foreach (var pair in initialValues) {
                 Guard.Require("initialValues", !String.IsNullOrEmpty(pair.Key), "Variable must have a name.");
 
                 Guard.Require("initialValues", pair.Value != null, "Variable '{0}': The type '{1}' is not supported.", pair.Key, pair.Value.GetType());
                 varNames[i] = pair.Key;
                 string varValue;
-                if (TessConvert.TryToString(pair.Value, out varValue))
-                {
+                if (TessConvert.TryToString(pair.Value, out varValue)) {
                     varValues[i] = varValue;
-                }
-                else
-                {
+                } else {
                     throw new ArgumentException(
                         String.Format("Variable '{0}': The type '{1}' is not supported.", pair.Key, pair.Value.GetType()),
                         "initialValues"
@@ -403,15 +557,11 @@ namespace Tesseract.Interop
         public static int BaseApiSetDebugVariable(HandleRef handle, string name, string value)
         {
             IntPtr valuePtr = IntPtr.Zero;
-            try
-            {
+            try {
                 valuePtr = MarshalHelper.StringToPtr(value, Encoding.UTF8);
                 return Native.BaseApiSetDebugVariable(handle, name, valuePtr);
-            }
-            finally
-            {
-                if (valuePtr != IntPtr.Zero)
-                {
+            } finally {
+                if (valuePtr != IntPtr.Zero) {
                     Marshal.FreeHGlobal(valuePtr);
                 }
             }
@@ -420,15 +570,11 @@ namespace Tesseract.Interop
         public static int BaseApiSetVariable(HandleRef handle, string name, string value)
         {
             IntPtr valuePtr = IntPtr.Zero;
-            try
-            {
+            try {
                 valuePtr = MarshalHelper.StringToPtr(value, Encoding.UTF8);
                 return Native.BaseApiSetVariable(handle, name, valuePtr);
-            }
-            finally
-            {
-                if (valuePtr != IntPtr.Zero)
-                {
+            } finally {
+                if (valuePtr != IntPtr.Zero) {
                     Marshal.FreeHGlobal(valuePtr);
                 }
             }
@@ -436,30 +582,38 @@ namespace Tesseract.Interop
 
         public static void Initialize()
         {
-            if (native == null)
-            {
+            if (native == null) {
                 LeptonicaApi.Initialize();
                 native = InteropRuntimeImplementer.CreateInstance<ITessApiSignatures>();
             }
         }
 
+        public static string ResultIteratorWordRecognitionLanguage(HandleRef handle)
+        {
+            // per docs (ltrresultiterator.h:118 as of 4897796 in github:tesseract-ocr/tesseract)
+            // this return value should *NOT* be deleted.
+            IntPtr txtHandle =
+                Native.ResultIteratorWordRecognitionLanguageInternal(handle);
+
+            return txtHandle != IntPtr.Zero
+                ? MarshalHelper.PtrToString(txtHandle, Encoding.UTF8)
+                : null;
+        }
+
         public static string ResultIteratorGetUTF8Text(HandleRef handle, PageIteratorLevel level)
         {
             IntPtr txtHandle = Native.ResultIteratorGetUTF8TextInternal(handle, level);
-            if (txtHandle != IntPtr.Zero)
-            {
+            if (txtHandle != IntPtr.Zero) {
                 var result = MarshalHelper.PtrToString(txtHandle, Encoding.UTF8);
                 TessApi.Native.DeleteText(txtHandle);
                 return result;
-            }
-            else
-            {
+            } else {
                 return null;
             }
         }
 
         /// <summary>
-        /// Returns the null terminated UTF-8 encoded text string for the current choice           
+        /// Returns the null terminated UTF-8 encoded text string for the current choice
         /// </summary>
         /// <remarks>
         /// NOTE: Unlike LTRResultIterator::GetUTF8Text, the return points to an
@@ -475,7 +629,5 @@ namespace Tesseract.Interop
         }
 
         // hOCR Extension
-
-
     }
 }