+++ src/Tesseract/Interop/Constants.cs
@@ -7,7 +7,11 @@ namespace Tesseract.Interop
     /// </summary>
     internal static class Constants
     {
-        public const string LeptonicaDllName = "liblept172";
-        public const string TesseractDllName = "libtesseract304";
+        public const string LeptonicaDllName = "leptonica-1.80.0";
+        public const string TesseractDllName = "tesseract41";
+        
+        // tesseract uses an int to represent true false values.
+        public const int TRUE = 1;
+        public const int FALSE = 0;
     }
 }