+++ src/Tesseract/PixConverter.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#if SYSTEM_DRAWING_SUPPORT
+
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Text;
@@ -34,3 +36,5 @@ namespace Tesseract
         }
     }
 }
+
+#endif